Move monodromy stuff in its own module
